Removed search from navigation to spec-sidebar
